+++ src/telebot.c
@@ -406,7 +406,7 @@ telebot_error_e telebot_free_webhook_info(telebot_webhook_info_t *info)
     return TELEBOT_ERROR_NONE;
 }
 
-telebot_error_e telebot_send_message(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_message(telebot_handler_t handle, long long int chat_id,
         char *text, char *parse_mode, bool disable_web_page_preview,
         bool disable_notification, int reply_to_message_id,
         const char *reply_markup)
@@ -427,8 +427,8 @@ telebot_error_e telebot_send_message(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_forward_message(telebot_handler_t handle, int chat_id,
-        int from_chat_id, bool disable_notification, int message_id)
+telebot_error_e telebot_forward_message(telebot_handler_t handle, long long int chat_id,
+        long long int from_chat_id, bool disable_notification, int message_id)
 {
     telebot_hdata_t * _handle = (telebot_hdata_t *)handle;
     if (_handle == NULL)
@@ -445,7 +445,7 @@ telebot_error_e telebot_forward_message(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_photo(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_photo(telebot_handler_t handle, long long int chat_id,
         char *photo, bool is_file, char *caption, bool disable_notification,
         int reply_to_message_id, char *reply_markup)
 {
@@ -464,7 +464,7 @@ telebot_error_e telebot_send_photo(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_audio(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_audio(telebot_handler_t handle, long long int chat_id,
         char *audio, bool is_file, int duration, char *performer, char *title,
         bool disable_notification, int reply_to_message_id, char *reply_markup)
 {
@@ -484,7 +484,7 @@ telebot_error_e telebot_send_audio(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_document(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_document(telebot_handler_t handle, long long int chat_id,
         char *document, bool is_file, bool disable_notification,
         int reply_to_message_id, char *reply_markup)
 {
@@ -503,7 +503,7 @@ telebot_error_e telebot_send_document(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_video(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_video(telebot_handler_t handle, long long int chat_id,
         char *video, bool is_file, int duration, int width, int height,
         char *caption, bool disable_notification, int reply_to_message_id,
         char *reply_markup)
@@ -524,7 +524,7 @@ telebot_error_e telebot_send_video(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_voice(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_voice(telebot_handler_t handle, long long int chat_id,
         char *voice, bool is_file, char *caption, int duration,
         bool disable_notification, int reply_to_message_id, char *reply_markup)
 {
@@ -543,7 +543,7 @@ telebot_error_e telebot_send_voice(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_video_note(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_video_note(telebot_handler_t handle, long long int chat_id,
         char *video_note, bool is_file, int duration, int length,
         bool disable_notification, int reply_to_message_id, char *reply_markup)
 {
@@ -563,7 +563,7 @@ telebot_error_e telebot_send_video_note(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_location(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_location(telebot_handler_t handle, long long int chat_id,
         float latitude, float longitude, bool disable_notification,
         int reply_to_message_id, char *reply_markup)
 {
@@ -582,7 +582,7 @@ telebot_error_e telebot_send_location(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_contact(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_contact(telebot_handler_t handle, long long int chat_id,
         char *phone_number, char *first_name, char *last_name,
         bool disable_notification, int reply_to_message_id, char *reply_markup)
 {
@@ -602,7 +602,7 @@ telebot_error_e telebot_send_contact(telebot_handler_t handle, int chat_id,
     return ret;
 }
 
-telebot_error_e telebot_send_chat_action(telebot_handler_t handle, int chat_id,
+telebot_error_e telebot_send_chat_action(telebot_handler_t handle, long long int chat_id,
         char *action)
 {
     telebot_hdata_t * _handle = (telebot_hdata_t *)handle;
@@ -736,7 +736,7 @@ telebot_error_e telebot_download_file(telebot_handler_t handle, char *file_id,
 }
 
 #if 0
-telebot_error_e telebot_delete_message(int chat_id, int message_id)
+telebot_error_e telebot_delete_message(long long int chat_id, int message_id)
 {
     if (g_handler == NULL)
         return TELEBOT_ERROR_NOT_SUPPORTED;
@@ -750,7 +750,7 @@ telebot_error_e telebot_delete_message(int chat_id, int message_id)
     return ret;
 }
 
-telebot_error_e telebot_send_sticker(int chat_id, char *sticker,
+telebot_error_e telebot_send_sticker(long long int chat_id, char *sticker,
         bool is_file, int reply_to_message_id, char *reply_markup)
 {
     if (g_handler == NULL)
